POND-19: adjust to inner outer logic
