+++ components/layout/LayoutSliderSlideInner.vue
@@ -1,0 +1,5 @@
+<template>
+    <swiper-slide class="flex relative max-w-full justify-center">
+        <slot />
+    </swiper-slide>
+</template>
